remove all 'as' casts

+++ gateway/src/error.rs
@@ -59,7 +59,7 @@ pub enum Error {
     #[error("invalid topic in query")]
     InvalidTopic,
     #[error("block range {0:?} not found in {1} parquet files")]
-    RangeNotFoundInParquetFiles((u32, u32), &'static str),
+    RangeNotFoundInParquetFiles((u64, u64), &'static str),
     #[error("too many toics in query. maximum is 4 but got {0}")]
     TooManyTopics(usize),
 }
